refactor blueprints (5/6): extract cockpit blueprint + url_for migration
Eleven cockpit-domain routes move to blueprints/cockpit.py:
- /publish (GET/POST), /login (GET/POST), /logout
- /scratch (GET/POST), /scratch/work (GET/POST)
- /cockpit/mode, /cockpit/mode/clear (recovered from sub-phase 4 — the
  range-extraction for ani routes accidentally pulled them along
  with the ani routes since they sat inside the same section
  visually; now correctly placed in cockpit blueprint and the
  ani.py copies removed)
- /cockpit/focus/break
- /cockpit/after-dark/{library,music,ani-loops}

WORK_MODE_PIN, AFTER_DARK_PIN, BRRR_WEBHOOK_URL, PASSWORD, and
SCRATCH_FILE redefined locally from os.environ.

Cross-blueprint dependency: imports ani_notify_publish from
blueprints.ani for the publish flow's notify hook.

`url_for` migrations triggered atomically:
- helpers/auth.py: url_for('login') → url_for('cockpit.login')
- blueprints/today.py: url_for('login') → url_for('cockpit.login')
- blueprints/below_deck.py: url_for('login') → url_for('cockpit.login')
- Inside cockpit blueprint, url_for('publish_status') → url_for('cockpit.publish_status'),
  url_for('login') → url_for('cockpit.login'), url_for('logout')
  unchanged (intra-blueprint).

Smoke-tested via Flask test_client: /cockpit/mode (wrong + correct
PINs both 200), /cockpit/mode/clear, /publish, /today/, /below-deck,
/ani/ping, /command-deck/ all 200.
